Add support for the ietf-yang-library model
Part of the restconf standard (RFC8040 section 10) says that a
restconf server must support the ietf-yang-library data model.
This model was most recently updated in RFC8525 and the the
actual ietf-yang-library model in revision 2019-01-04.

Currently there is no way to support import-only modules as this
information is not currently stored in the XML models used by
restconf.