Use a new method to get loaded model information
A problem with the current method for getting model information was
found when two models with a duplicate top schema node were loaded.
This change uses a list of model information stored in the
apteryx-xml library. The list is created at start time when the
models were loaded.

Fixed up the test_restconf_yang_library_tree as it was missing one
of the models.