Support translated YANG models.
This change allows one model to be translated to another model. This is
useful if multiple YANG models for the same type of data exist. In this
case a back-end super YANG model is created which is a combination of
the contributing models, and requests to the individual models are
translated into a request of the super model and responses are
translated back into the individual requesting models.