Allow the user to limit the models loaded by apteryx-rest
A new startup option "-n" has been added so the user can specify a file
containing a list of supported models. All other models in the schema
directory, not listed in the supported models file, will not be loaded.
If the option "-n" is not present in the command line, then all models will
be loaded.