Handle list keys with a '/' in the value
It is valid to have a / in the key value, but must be
encoded in rest URI's to avoid being interpreted as a
path element. As you are not allowed to encode normal
path elements using %2F we can avoid decoding it and
hence keep it encoded while processing. We store the
encoded value as part of the apteryx path.

Add tests for / in the key as well as other reserved
characters.