Add support for testing appweb as server
Requires 2 small patches to appweb-8.2.3
1. Appweb fastcgi handler assumes if the location header is present
that the app wants the client to redirect. In restconf the location
header is used to indicate the location of the resource. So we need
to squash the redirect support.
2. The vanilla appweb code thinks it is being clever by sending the
last \r\n of chunked data with the start of the next block, but this
is not compliant with SSE. So fix the chunking code to send the
proper header and trailer <SIZE>\r\n<DATA>\r\n.