Fix tests based on changes to schemas
Recent changes to schemas in apteryx-xml caused a number of tests to
fail. This change restores all tests to passing status.

The tests in test_query.py had to be modified because the when clause
on the node toys was be acted on due to the new default value assigned
to config/type.