Get latest timestamp for last-modified http header field
If a child on the path is a refresher, the timestamp set
in the last-modified http header field will not be upto
date. This is because only refreshers matching the timestamp
are triggered, not children.

Get the latest timestamp to store in last-modified http header
field after the data has been fetched via apteryx_query().