Update fcgi to 2.4.6
Update to bring in upstream fix to CVE-2025-23016.

The original upstream has been archived, and this new version is where
it is maintained (as "fcgi2").

With -Werror, the redefine of PACKAGE_NAME, _VERSION, _TARNAME, and
_STRING was causing a compilation issue. Added #undef statements just
before include of fcgi_config.h to prevent the clash.

Updated run.sh with new FCGI URL and autoreconf build process.